Assets config changes, Import script stability points fixed
